pyproject.toml
[tool.poetry]
name = "task-management-api"
version = "0.1.0"
description = ""
authors = ["Isaias Valle <guitartsword@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
sqlmodel = "^0.0.19"
psycopg = {extras = ["binary"], version = "^3.2.1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
# Pin bcrypt until passlib supports the latest
bcrypt = "4.0.1"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
